#805 moving Screening Status from the Taxonomies to the Apply admin
